aws_lambda/requirements/base.txt
python-twitter==3.5
requests==2.25.1